chore: migrate to json2jsii
Extract the jsii struct generator into a new module called json2jsii and refactor the CLI to use it.
