+++ packages/cdk8s-cli/lib/import/k8s.ts
@@ -1,10 +1,11 @@
 import { CodeMaker } from 'codemaker';
 import { JSONSchema4 } from 'json-schema';
-import { TypeGenerator } from './type-generator';
 import { ImportBase } from './base';
 import { ApiObjectName, parseApiTypeName, compareApiVersions } from './k8s-util';
 import { ImportSpec } from '../config';
 import { download } from '../util';
+import { TypeGenerator } from 'json2jsii';
+import { generateConstruct } from './codegen';
 
 const DEFAULT_API_VERSION = '1.15.0';
 
@@ -59,23 +60,27 @@ export class ImportKubernetesApi extends ImportBase {
   
     const topLevelObjects = selectApiObjects(map, { include: this.options.include });
   
-    code.line(`// generated by cdk8s`);
-    code.line(`import { ApiObject } from 'cdk8s';`);
-    code.line(`import { Construct } from 'constructs';`);
-    code.line();
   
-    const typeGenerator = new TypeGenerator(schema, { exclude: this.options.exclude });
+    const typeGenerator = new TypeGenerator({
+      definitions: schema.definitions,
+      exclude: this.options.exclude
+    });
   
     for (const o of topLevelObjects) {
       this.emitConstructForApiObject(typeGenerator, o);
     }
   
-    typeGenerator.generate(code);
+    code.line(`// generated by cdk8s`);
+    code.line(`import { ApiObject } from 'cdk8s';`);
+    code.line(`import { Construct } from 'constructs';`);
+    code.line();
+
+    code.line(typeGenerator.render());
   }
 
   private emitConstructForApiObject(typeGenerator: TypeGenerator, apidef: ApiObjectDefinition) {
     const objectName = getObjectName(apidef);
-    typeGenerator.emitConstruct({
+    generateConstruct(typeGenerator, {
       fqn: apidef.fullname,
       group: objectName.group,
       kind: objectName.kind,
