+++ packages/cdk8s-cli/package.json
@@ -64,7 +64,7 @@
   "devDependencies": {
     "@types/fs-extra": "^8.1.0",
     "@types/jest": "^25.1.2",
-    "@types/json-schema": "^7.0.4",
+    "@types/json-schema": "^7.0.5",
     "@typescript-eslint/eslint-plugin": "^2.20.0",
     "@typescript-eslint/parser": "^2.20.0",
     "eslint": "^6.8.0",
